Remove all of SecureSocial from common
