+++ kifi-backend/modules/common/conf/commonService.routes
@@ -23,13 +23,8 @@ GET     /up                                   @com.keepit.common.admin.ServiceCo
 GET     /internal/benchmark                   @com.keepit.common.healthcheck.CommonBenchmarkController.benchmarksResults()
 GET     /internal/version                     @com.keepit.common.healthcheck.CommonBenchmarkController.version()
 GET     /internal/clusters/refresh            @com.keepit.common.admin.ServiceController.forceRefresh()
-GET		  /internal/clusters/mystatus			      @com.keepit.common.admin.ServiceController.myStatus
-GET		  /internal/clusters/topology			      @com.keepit.common.admin.ServiceController.topology
+GET		/internal/clusters/mystatus			  @com.keepit.common.admin.ServiceController.myStatus
+GET		/internal/clusters/topology			  @com.keepit.common.admin.ServiceController.topology
 GET     /internal/common/threadSummary        @com.keepit.common.admin.ServiceController.threadSummary
 GET     /internal/common/threadDetails        @com.keepit.common.admin.ServiceController.threadDetails(name: String ?= "", state: String ?= "", stack: String ?= "", sort: String ?= "")
-GET     /internal/cache/removeAll                 @com.keepit.common.admin.ServiceController.removeAllFromLocalCache(prefix: Option[String] ?= None)
-
-
-# OAuth provider entry points (e.g. "facebook")
-GET     /authenticate/:provider     com.keepit.social.providers.ProviderController.authenticate(provider)
-POST    /authenticate/:provider     com.keepit.social.providers.ProviderController.authenticateByPost(provider)
+GET     /internal/cache/removeAll             @com.keepit.common.admin.ServiceController.removeAllFromLocalCache(prefix: Option[String] ?= None)
