+++ kifi-backend/modules/graph/app/com/keepit/graph/GraphModule.scala
@@ -8,7 +8,6 @@ import com.keepit.abook.ProdABookServiceClientModule
 import com.keepit.graph.manager.{ GraphManagerPluginModule, GraphManagerModule }
 import com.keepit.graph.common.store.GraphStoreModule
 import com.keepit.graph.common.cache.GraphCacheModule
-import com.keepit.social.RemoteSecureSocialModule
 import com.keepit.cortex.ProdCortexServiceClientModule
 import com.keepit.common.zookeeper.ServiceTypeModule
 import com.keepit.common.service.ServiceType
@@ -35,5 +34,4 @@ trait GraphModule extends ConfigurationModule with CommonServiceModule {
   val elizaServiceClientModule = ProdElizaServiceClientModule()
   val abookServiceClientModule = ProdABookServiceClientModule()
   val cortexServiceClientModule = ProdCortexServiceClientModule()
-  val secureSocialModule = RemoteSecureSocialModule()
 }
