+++ kifi-backend/modules/search/test/com/keepit/search/controllers/ext/ExtSearchControllerTest.scala
@@ -7,7 +7,7 @@ import org.specs2.mutable._
 import com.keepit.model._
 import com.keepit.common.db.{ ExternalId }
 import com.keepit.common.actor.FakeActorSystemModule
-import com.keepit.common.controller.{ FakeSecureSocialClientIdModule, FakeUserActionsHelper, FakeUserActionsModule }
+import com.keepit.common.controller.{ FakeUserActionsHelper, FakeUserActionsModule }
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import com.keepit.search._
@@ -19,7 +19,6 @@ class ExtSearchControllerTest extends Specification with SearchTestInjector {
   def modules = {
     Seq(
       FakeActorSystemModule(),
-      FakeSecureSocialClientIdModule(),
       FakeUserActionsModule(),
       FixedResultIndexModule(),
       PlayAppConfigurationModule(),
