+++ kifi-backend/modules/shoebox/app/com/keepit/common/oauth/RichIdentity.scala
