+++ kifi-backend/modules/shoebox/conf/play.plugins
@@ -1,3 +1,9 @@
+9994:com.keepit.social.SecureSocialAuthenticatorStore
+9995:com.keepit.social.SecureSocialIdGenerator
+9996:securesocial.core.providers.utils.DefaultPasswordValidator
+9997:com.keepit.social.SecureSocialTemplatesPlugin
+9998:com.keepit.social.SecureSocialUserService
+9999:securesocial.core.providers.utils.BCryptPasswordHasher
 10001:com.keepit.social.providers.FacebookProvider
 10002:com.keepit.social.providers.TwitterProvider
 10003:com.keepit.social.providers.LinkedInProvider
